Removed new_policy and merged new code into policy
- ignore error on windows when trying to delete test sqlite file
