+++ packages/syft/src/syft/core/node/new/policy.py
@@ -24,6 +24,7 @@ from result import Ok
 from result import Result
 
 # relative
+from .action_object import ActionObject
 from .api import NodeView
 from .context import AuthedServiceContext
 from .context import NodeServiceContext
@@ -43,7 +44,9 @@ from .syft_object import SyftObject
 from .transforms import TransformContext
 from .transforms import generate_id
 from .transforms import transform
+from .twin_object import TwinObject
 from .uid import UID
+from .unparse import unparse
 
 PolicyUserVerifyKeyPartitionKey = PartitionKey(
     key="user_verify_key", type_=SyftVerifyKey
@@ -51,9 +54,54 @@ PolicyUserVerifyKeyPartitionKey = PartitionKey(
 PyCodeObject = Any
 
 
+def extract_uid(v: Any) -> UID:
+    value = v
+    if isinstance(v, ActionObject):
+        value = v.id
+    if isinstance(v, TwinObject):
+        value = v.id
+
+    if not isinstance(value, UID):
+        raise Exception(f"Input {v} must have a UID not {type(v)}")
+    return value
+
+
+def filter_only_uids(results: Any) -> Dict[str, Any]:
+    if not hasattr(results, "__len__"):
+        results = [results]
+
+    if isinstance(results, list):
+        output_list = []
+        for v in results:
+            output_list.append(extract_uid(v))
+        return output_list
+    elif isinstance(results, dict):
+        output_dict = {}
+        for k, v in results.items():
+            output_dict[k] = extract_uid(v)
+        return output_dict
+    return extract_uid(results)
+
+
 class Policy(SyftObject):
+    # version
     __canonical_name__ = "Policy"
     __version__ = SYFT_OBJECT_VERSION_1
+
+    id: UID
+
+    @property
+    def policy_code(self) -> str:
+        cls = type(self)
+        op_code = inspect.getsource(cls)
+        return op_code
+
+    def public_state() -> None:
+        raise NotImplementedError
+
+    @property
+    def valid(self) -> Union[SyftSuccess, SyftError]:
+        return SyftSuccess(message="Policy is valid.")
 
 
 @serializable()
@@ -110,50 +158,36 @@ class InputPolicyState(SyftObject):
     __version__ = SYFT_OBJECT_VERSION_1
 
 
-class InputPolicy(SyftObject):
+class InputPolicy(Policy):
     # version
     __canonical_name__ = "InputPolicy"
     __version__ = SYFT_OBJECT_VERSION_1
 
-    # relative
-    from .api import NodeView
-
     id: UID
     inputs: Dict[NodeView, Any]
     node_uid: Optional[UID]
-    state_type: Optional[Type[InputPolicyState]]
 
     def __init__(self, *args: Any, **kwargs: Any) -> None:
         # TODO: This method initialization would conflict if one of the input variables
         # to the code submission function happens to be id or inputs
         uid = UID()
         node_uid = None
-        state_type = None
         if "id" in kwargs:
             uid = kwargs["id"]
         if "node_uid" in kwargs:
             node_uid = kwargs["node_uid"]
-        if "state_type" in kwargs:
-            state_type = kwargs["state_type"]
 
         # finally get inputs
         if "inputs" in kwargs:
             kwargs = kwargs["inputs"]
         else:
             kwargs = partition_by_node(kwargs)
-        super().__init__(
-            id=uid, inputs=kwargs, node_uid=node_uid, state_type=state_type
-        )
+        super().__init__(id=uid, inputs=kwargs, node_uid=node_uid)
 
-    def filter_kwargs(kwargs: Dict[str, Any]) -> Dict[str, Any]:
+    def filter_kwargs(
+        self, kwargs: Dict[str, Any], context: AuthedServiceContext, code_item_id: UID
+    ) -> Dict[str, Any]:
         raise NotImplementedError
-
-    # def __getitem__(self, key: Union[int, str]) -> Optional[SyftObject]:
-    #     if isinstance(key, int):
-    #         key = list(self.inputs.keys())[key]
-    #     uid = self.inputs[key]
-    #     # TODO Add NODE UID or LINK so we can resolve this object
-    #     return uid
 
     @property
     def assets(self) -> List[Asset]:
@@ -347,45 +381,117 @@ class OutputPolicyStateExecuteOnce(OutputPolicyStateExecuteCount):
     limit: int = 1
 
 
-class OutputPolicy(SyftObject):
+class OutputPolicy(Policy):
     # version
     __canonical_name__ = "OutputPolicy"
     __version__ = SYFT_OBJECT_VERSION_1
 
-    id: UID
+    output_history: List[OutputHistory] = []
     outputs: List[str] = []
-    state_type: Optional[Type[OutputPolicyState]]
     node_uid: Optional[UID]
 
-    def update() -> None:
-        raise NotImplementedError
+    def apply_output(
+        self,
+        context: NodeServiceContext,
+        outputs: Any,
+    ) -> Any:
+        output_uids = filter_only_uids(outputs)
+        if isinstance(output_uids, UID):
+            output_uids = [output_uids]
+        history = OutputHistory(
+            output_time=DateTime.now(),
+            outputs=output_uids,
+            executing_user_verify_key=context.credentials,
+        )
+        self.output_history.append(history)
+        return outputs
+
+
+@serializable()
+class OutputPolicyExecuteCount(OutputPolicy):
+    __canonical_name__ = "OutputPolicyExecuteCount"
+    __version__ = SYFT_OBJECT_VERSION_1
+
+    count: int = 0
+    limit: int
+
+    def apply_output(
+        self,
+        context: NodeServiceContext,
+        outputs: Any,
+    ) -> Optional[Any]:
+        if self.count < self.limit:
+            super().apply_output(context, outputs)
+            self.count += 1
+            return outputs
+        return None
 
     @property
-    def policy_code(self) -> str:
-        cls = type(self)
-        op_code = inspect.getsource(cls)
-        if self.state_type:
-            state_code = inspect.getsource(self.state_type)
-            op_code += "\n" + state_code
-        return op_code
+    def valid(self) -> Union[SyftSuccess, SyftError]:
+        is_valid = self.count < self.limit
+        if is_valid:
+            return SyftSuccess(
+                message=f"Policy is still valid. count: {self.count} < limit: {self.limit}"
+            )
+        return SyftError(
+            message=f"Policy is no longer valid. count: {self.count} >= limit: {self.limit}"
+        )
+
+    def public_state(self) -> None:
+        return {"limit": self.limit, "count": self.count}
 
 
-class CustomOutputPolicy(OutputPolicy):
+@serializable()
+class OutputPolicyExecuteOnce(OutputPolicyExecuteCount):
+    __canonical_name__ = "OutputPolicyExecuteOnce"
+    __version__ = SYFT_OBJECT_VERSION_1
+
+    limit: int = 1
+
+
+@serializable()
+class SingleExecutionExactOutput(OutputPolicyExecuteCount):
+    # version
+    __canonical_name__ = "SingleExecutionExactOutput"
+    __version__ = SYFT_OBJECT_VERSION_1
+
+    limit: int = 1
+
+
+class CustomPolicy(Policy):
+    # version
+    __canonical_name__ = "CustomPolicy"
+    __version__ = SYFT_OBJECT_VERSION_1
+
+    init_args: Dict[str, Any] = {}
+    init_kwargs: Dict[str, Any] = {}
+
+    def __init__(self, *args, **kwargs) -> None:
+        super().__init__(init_args=args, init_kwargs=kwargs, *args, **kwargs)
+
+
+class CustomOutputPolicy(CustomPolicy, OutputPolicy):
     # version
     __canonical_name__ = "CustomOutputPolicy"
     __version__ = SYFT_OBJECT_VERSION_1
-    init_args: Dict[str, Any] = {}
-    kwargs: Dict[str, Any] = {}
 
-    def __init__(self, *args, **kwargs) -> None:
-        super().__init__(**kwargs)
-        self.init_args = args
-        self.kwargs = kwargs
+    def apply_output(
+        self,
+        context: NodeServiceContext,
+        outputs: Any,
+    ) -> Optional[Any]:
+        return outputs
+
+
+class CustomInputPolicy(CustomPolicy, InputPolicy):
+    # version
+    __canonical_name__ = "CustomInputPolicy"
+    __version__ = SYFT_OBJECT_VERSION_1
 
 
 @serializable()
 class UserPolicy(Policy):
-    __canonical_name__ = "UserPolicy"
+    __canonical_name__ = "UserPolicy_2"
     __version__ = SYFT_OBJECT_VERSION_1
 
     id: UID
@@ -400,14 +506,22 @@ class UserPolicy(Policy):
     byte_code: PyCodeObject
     status: UserPolicyStatus = UserPolicyStatus.SUBMITTED
     state_type: Optional[Type] = None
+    policy_version: int
 
     @property
     def byte_code(self) -> Optional[PyCodeObject]:
         return compile_byte_code(self.parsed_code)
 
     @property
-    def valid(self) -> Union[SyftSuccess, SyftError]:
-        return SyftSuccess(message="Policy is valid.")
+    def policy_code(self) -> str:
+        return self.raw_code
+
+    def apply_output(
+        self,
+        context: NodeServiceContext,
+        outputs: Any,
+    ) -> Optional[Any]:
+        return outputs
 
 
 @serializable()
@@ -434,9 +548,10 @@ def hash_code(context: TransformContext) -> TransformContext:
 
 
 def generate_unique_class_name(context: TransformContext) -> TransformContext:
+    # TODO: Do we need to check if the initial name contains underscores?
     code_hash = context.output["code_hash"]
     service_class_name = context.output["class_name"]
-    unique_name = f"user_func_{service_class_name}_{context.credentials}_{code_hash}"
+    unique_name = f"{service_class_name}_{context.credentials}_{code_hash}"
     context.output["unique_name"] = unique_name
     return context
 
@@ -455,8 +570,43 @@ def process_class_code(raw_code: str, class_name: str, input_kwargs: List[str]) 
     v = GlobalsVisitor()
     v.visit(tree)
 
-    f = tree.body[0]
-    f.decorator_list = []
+    if len(tree.body) != 1 or not isinstance(tree.body[0], ast.ClassDef):
+        raise Exception(
+            "Class code should only contain the Class Definition for your policy."
+        )
+
+    old_class = tree.body[0]
+    if len(old_class.bases) != 1 or old_class.bases[0].id not in [
+        "CustomInputPolicy",
+        "CustomOutputPolicy",
+    ]:
+        raise Exception(
+            "Class code should either implement CustomInputPolicy or CustomOutputPolicy"
+        )
+
+    # TODO: changes the bases
+
+    serializable_name = ast.Name(id="serializable", ctx=ast.Load())
+    serializable_decorator = ast.Call(
+        func=serializable_name,
+        args=[],
+        keywords=[ast.keyword(arg="recursive_serde", value=ast.Constant(value=True))],
+    )
+
+    new_class = tree.body[0]
+    version = None
+    for stmt in new_class.body:
+        if isinstance(stmt, ast.Assign):
+            if stmt.targets[0].id == "__version__":
+                version = stmt.value.value
+                break
+    if version is None:
+        raise Exception(
+            "Version cannot be found. Please specify it in your custom policy"
+        )
+    new_class.name = class_name
+    new_class.decorator_list = [serializable_decorator]
+    return unparse(new_class), version
 
 
 def check_class_code(context: TransformContext) -> TransformContext:
@@ -466,13 +616,13 @@ def check_class_code(context: TransformContext) -> TransformContext:
     # parse init signature
     # check dangerous libraries, maybe compile_restricted already does that
     try:
-        processed_code = process_class_code(
-            raw_code=context.output["code"],
+        processed_code, policy_version = process_class_code(
+            raw_code=context.output["raw_code"],
             class_name=context.output["unique_name"],
             input_kwargs=context.output["input_kwargs"],
         )
         context.output["parsed_code"] = processed_code
-
+        context.output["policy_version"] = policy_version
     except Exception as e:
         raise e
     return context
@@ -506,17 +656,16 @@ def generate_signature(context: TransformContext) -> TransformContext:
     return context
 
 
-def serialization_addon(context: TransformContext) -> TransformContext:
-    policy_code = context.output["code"]
-    context.output["code"] = "@serializable()\n" + policy_code
-    return context
+# def serialization_addon(context: TransformContext) -> TransformContext:
+#     policy_code = context.output["code"]
+#     context.output["code"] = "@serializable()\n" + policy_code
+#     return context
 
 
 @transform(SubmitUserPolicy, UserPolicy)
 def submit_policy_code_to_user_code() -> List[Callable]:
     return [
         generate_id,
-        serialization_addon,
         hash_code,
         generate_unique_class_name,
         generate_signature,
@@ -539,8 +688,13 @@ def execute_policy_code(user_policy: UserPolicy):
         # syft absolute
         import syft as sy  # noqa: F401 # provide sy.Things to user code
 
-        exec(user_policy.byte_code)  # nosec
-        policy_class = eval(user_policy.class_name)  # nosec
+        class_name = f"{user_policy.unique_name}_{user_policy.policy_version}"
+        print("Exec context", file=stderr_)
+        if class_name in user_policy.__object_version_registry__.keys():
+            policy_class = user_policy.__object_version_registry__[class_name]
+        else:
+            exec(user_policy.byte_code)  # nosec
+            policy_class = eval(user_policy.unique_name)  # nosec
 
         sys.stdout = stdout_
         sys.stderr = stderr_
@@ -549,28 +703,6 @@ def execute_policy_code(user_policy: UserPolicy):
 
     except Exception as e:
         print("execute_byte_code failed", e, file=stderr_)
-        try:
-            stdout = StringIO()
-            stderr = StringIO()
-
-            sys.stdout = stdout
-            sys.stderr = stderr
-            # exec(user_policy.byte_code)  # nosec
-            # policy_class = eval(user_policy.class_name)  # nosec
-            print(
-                user_policy.__object_version_registry__["RepeatedCallPolicy_1"],
-                file=stderr_,
-            )
-            policy_class = user_policy.__object_version_registry__[
-                "RepeatedCallPolicy_1"
-            ]
-
-            sys.stdout = stdout_
-            sys.stderr = stderr_
-
-            return policy_class
-        except Exception as e:
-            print("execute_byte_code failed", e, file=stderr_)
 
     finally:
         sys.stdout = stdout_
