+++ packages/syft/src/syft/core/node/new/user_code.py
@@ -29,17 +29,17 @@ from .context import AuthedServiceContext
 from .credentials import SyftVerifyKey
 from .dataset import Asset
 from .document_store import PartitionKey
-from .new_policy import CustomInputPolicy
-from .new_policy import CustomOutputPolicy
-from .new_policy import InputPolicy
-from .new_policy import OutputPolicy
-from .new_policy import Policy
-from .new_policy import SubmitUserPolicy
-from .new_policy import UserPolicy
 from .node import NodeType
 from .node_metadata import EnclaveMetadata
+from .policy import CustomInputPolicy
+from .policy import CustomOutputPolicy
+from .policy import InputPolicy
 from .policy import InputPolicyState
+from .policy import OutputPolicy
 from .policy import OutputPolicyState
+from .policy import Policy
+from .policy import SubmitUserPolicy
+from .policy import UserPolicy
 from .policy_service import PolicyService
 from .serializable import serializable
 from .syft_object import SYFT_OBJECT_VERSION_1
