+++ packages/syft/src/syft/core/node/new/user_code_service.py
@@ -12,14 +12,14 @@ from ....telemetry import instrument
 from .context import AuthedServiceContext
 from .document_store import DocumentStore
 from .linked_obj import LinkedObject
-from .new_policy import OutputHistory
-from .new_policy import OutputPolicy
-from .new_policy import SubmitUserPolicy
-from .new_policy import UserPolicy
-from .new_policy import UserPolicyStatus
-from .new_policy import get_policy_object
-from .new_policy import init_policy
-from .new_policy import update_policy_state
+from .policy import OutputHistory
+from .policy import OutputPolicy
+from .policy import SubmitUserPolicy
+from .policy import UserPolicy
+from .policy import UserPolicyStatus
+from .policy import get_policy_object
+from .policy import init_policy
+from .policy import update_policy_state
 from .request import EnumMutation
 from .request import UserCodeStatusChange
 from .response import SyftError
